requirements.txt
requests==2.32.5
pandas==2.3.2
python-dotenv==1.1.1
beautifulsoup4==4.13.4
playsound==1.2.2
pathspec==0.12.1
tqdm==4.67.1
pyhumps==3.8.0
